requirements_auth_service.txt
fastapi>=0.95.0
uvicorn>=0.21.1
asyncpg>=0.27.0
python-dotenv>=1.0.0
pycryptodome>=3.17.0
pydantic>=2.0.0
